# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Oficina SOAT",
        "description": "Documentação da api do tech challenge FIAP 06.2025",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/peca_insumo": {
            "get": {
                "tags": [
                    "PecaInsumo"
                ],
                "summary": "Listar peças e insumos",
                "description": "Faz a listagem paginada de todas as peças e insumos cadastrados",
                "operationId": "d41606b9e20da90731da60f27d7d6245",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página para paginação",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de peças e insumos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "gtin": {
                                                        "type": "string",
                                                        "example": "7891234567890"
                                                    },
                                                    "descricao": {
                                                        "type": "string",
                                                        "example": "Filtro de óleo"
                                                    },
                                                    "valor_custo": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 25.5
                                                    },
                                                    "valor_venda": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 45.9
                                                    },
                                                    "qtd_atual": {
                                                        "type": "integer",
                                                        "example": 100
                                                    },
                                                    "qtd_segregada": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "ativo"
                                                    },
                                                    "excluido": {
                                                        "type": "boolean",
                                                        "example": false
                                                    },
                                                    "data_cadastro": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-01-15T10:30:00.000000Z"
                                                    },
                                                    "data_atualizacao": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "data_exclusao": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": null,
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8080/api/peca_insumo?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8080/api/peca_insumo?page=5"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8080/api/peca_insumo?page=2",
                                            "nullable": true
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://localhost:8080/api/peca_insumo"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 75
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro interno do servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "PecaInsumo"
                ],
                "summary": "Cadastrar peça ou insumo",
                "description": "Cadastra uma nova peça ou insumo",
                "operationId": "79508ea81b195de084ebcfe24514aba3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "gtin",
                                    "descricao",
                                    "valor_custo",
                                    "valor_venda",
                                    "qtd_atual"
                                ],
                                "properties": {
                                    "gtin": {
                                        "type": "string",
                                        "example": "7891234567890"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Filtro de óleo"
                                    },
                                    "valor_custo": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25.5
                                    },
                                    "valor_venda": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 45.9
                                    },
                                    "qtd_atual": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "qtd_segregada": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "ativo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Peça ou insumo cadastrado com sucesso"
                    },
                    "422": {
                        "description": "Dados de entrada inválidos"
                    }
                }
            }
        },
        "/api/peca_insumo/{id}": {
            "get": {
                "tags": [
                    "PecaInsumo"
                ],
                "summary": "Obter peça ou insumo por ID",
                "description": "Obtém uma peça ou insumo específico pelo seu ID",
                "operationId": "cefa90f9b4400f9fcedeb1eb63520abc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Peça ou insumo encontrado com sucesso"
                    },
                    "404": {
                        "description": "Peça ou insumo não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "PecaInsumo"
                ],
                "summary": "Atualizar peça ou insumo",
                "description": "Atualiza os dados de uma peça ou insumo pelo seu ID",
                "operationId": "e56262ad8e57187b0b4f1aa5080c84e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "gtin": {
                                        "type": "string",
                                        "example": "7891234567890"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Filtro de óleo atualizado"
                                    },
                                    "valor_custo": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 30
                                    },
                                    "valor_venda": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "qtd_atual": {
                                        "type": "integer",
                                        "example": 85
                                    },
                                    "qtd_segregada": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "ativo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Peça ou insumo atualizado com sucesso"
                    },
                    "404": {
                        "description": "Peça ou insumo não encontrado"
                    },
                    "422": {
                        "description": "Dados de entrada inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "PecaInsumo"
                ],
                "summary": "Remover peça ou insumo",
                "description": "Remove uma peça ou insumo pelo seu ID",
                "operationId": "81682879907838e6d8dccd9bb42249e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da peça ou insumo a ser removido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Peça ou insumo removido com sucesso (sem conteúdo de retorno)"
                    },
                    "404": {
                        "description": "Peça ou insumo não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Peça ou insumo não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro interno do servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}